handle invalid access token error (#49)
